refactor(backend): split farm/models.py into a domain-organized models package
Break the 1656-line farm/models.py into farm/models/ with one module per
domain (base, projects, structure, cultures, planning, notes, history),
all re-exported from models/__init__.py so every 'from farm.models import
X' keeps working unchanged.

Migration state is byte-identical (makemigrations --check reports no
changes): the two upload-path callables stay defined in models/__init__.py
so their serialized 'farm.models.note_attachment_upload_path' path in
migration 0023 remains valid, and moving model classes between modules
does not change their (app_label, model_name) identity. No FK
declarations changed — the four existing string-ref FKs already break the
Culture<->PublicCulture and history->Culture cycles, so the domain modules
import in a clean dependency order.

All 382 backend tests pass. Docs updated to point at the new package.
