+++ FIELD_BED_AREA_IMPLEMENTATION.md
@@ -3,7 +3,7 @@
 ## Overview
 This document summarizes the implementation of area-only storage for Field and Bed entities in v1, preparing for future polygon/geometry support.
 
-> **Status note (2026-07):** the model/validation details below are still current (see `backend/farm/models.py`). The frontend section references `frontend/src/pages/Beds.tsx`, which no longer exists as a standalone page — field/bed management now lives in `FieldsBedsPage.tsx`/`FieldsBedsHierarchy.tsx`/`GraphicalFields.tsx`. See [`docs/data-model.md`](./docs/data-model.md#2-farm-structure-locations--fields--beds) for the current farm-structure model overview.
+> **Status note (2026-07):** the model/validation details below are still current (see `backend/farm/models/structure.py`). The frontend section references `frontend/src/pages/Beds.tsx`, which no longer exists as a standalone page — field/bed management now lives in `FieldsBedsPage.tsx`/`FieldsBedsHierarchy.tsx`/`GraphicalFields.tsx`. See [`docs/data-model.md`](./docs/data-model.md#2-farm-structure-locations--fields--beds) for the current farm-structure model overview.
 
 ## Issue Requirements
 - ✅ Only the area (sqm) is stored in v1
@@ -15,7 +15,7 @@ This document summarizes the implementation of area-only storage for Field and B
 
 ### Backend Changes
 
-#### Models (backend/farm/models.py)
+#### Models (backend/farm/models/structure.py)
 - **Field Model**: Already had `area_sqm` field (DecimalField, max_digits=10, decimal_places=2)
   - Added validation: MIN_AREA_SQM = 0.01 sqm, MAX_AREA_SQM = 1,000,000 sqm (100 hectares)
   - Implemented `clean()` method for custom validation
