+++ backend/farm/models/base.py
@@ -1,0 +1,14 @@
+"""Shared abstract base model."""
+
+
+from django.db import models
+
+
+class TimestampedModel(models.Model):
+    """Abstract base model with created/updated timestamps."""
+
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        abstract = True
